requirements.txt
torchvision
Pillow
